Stop motor on error, code cleanup
I introduced macros for logging to serial, SD-card,
and both.
This greatly simplifies and shortens the code.

If you want to test the macros, copy their definitions
to a test.c file, add some lines using the macros,
and run gcc on it like this:
   gcc -E test.c
The -E option stops after the C-preprocessor (macros,
ifdef etc.) and shows the result.